MDL-41273 editpdf: Fix focus jumping on delete comment
When an empty comment is blurred it is deleted - this was redrawing
the entire page - which caused the focus to shift. So if you created
an empty comment - then clicked in a real comment, the focus would shift
to a different comment. I fixed this by not redrawing the entire page -
only removing the drawable related to the comment.