MDL-42430 Atto: Accessibility improvements
Move the title attribute to the button containing the image.
This fixes a warning from the chrome accessibility inspector, and shows
tooltips in firefox.

As part of this change, we are generating the full image tags in js - so we pass much
less data from php through to the js.

Also - alpha sort lang strings and remove redundant strings.