MDL-43869 Atto: Fix broken repositories.
The previous change on this issue accidentally fixed a selector,
so the repository API was recieving a valid node instead of null.
This means that it calls it's undocumented and dodgy callbacks on the node,
which are really calls directly to the TinyMCE API (just well buried).

But - by not setting the editor_target property, the api will not do
the secret handshakes.