MDL-43855 Atto: Add an equation editor
This equation editor relies on whatever the currently configured filter is to do the rendering.
If there is no filter that handles latex ($$ blah $$) - this plugin will not show up.

This will not work with the solutions on the forums of adding MathJax in the header of the page,
but it will work with a real mathjax filter written for Moodle (work in progress). It works with
the existing tex filter.