MDL-50040 Output: Skip the subtype when searching for templates
The subtype is too specific when searching for mustache templates and makes it
harder to determine the correct folder when overriding a template in a theme. It
would also mean we would need to prevent renderer subtypes from using templates from
another subtype in the same plugin (or have exhaustive search paths).