MDL-51963 tool_lp: Really add the scaleid field on competency
Still without any intention to use it, we add the scale field in
the database just so we can already write the right validation for
the scale ID which is pretty important as our code will start relying
on it. E.g. we only support scaleid null, or greater than 0. Adding
the validation method clearly documents what our expectations are.