SETTINGS navigation: Add firstsubmenu nav element
Pull the first sub-menu out of the settings menu and make it it's own thing in the page.

Re-order settings sub-nav nodes so there is a backup branch that will mostly be the thing that is pulled in to the page.

Add a few obvious page buttons here and there.

I don't think this is the final solution.